backend/requirements.txt
# Second Brain Backend - Python Dependencies
# 
# Install with: pip install -r requirements.txt
# 
# Note: For complete environment setup including system dependencies,
#       use the project's environment.yml with miniconda instead.

# ==========================================
# Web Framework & API
# ==========================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0

# ==========================================
# Database & ORM
# ==========================================
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
greenlet>=3.0.0  # Required for SQLAlchemy async support
psycopg2-binary>=2.9.9
neo4j>=5.16.0
redis>=5.0.1

# ==========================================
# Task Queue & Background Processing
# ==========================================
celery>=5.3.6
flower>=2.0.1
apscheduler>=3.10.0          # Scheduled job runner

# ==========================================
# LLM Integration & Processing
# ==========================================
litellm>=1.40.0
openai>=1.12.0
anthropic>=0.18.0
tiktoken>=0.7.0
tenacity>=8.2.3
aisuite
langchain-text-splitters>=0.2.0  # Text chunking for RAG pipelines

# ==========================================
# OCR & Document Processing
# ==========================================
mistralai>=0.1.0
pillow>=10.2.0
pymupdf>=1.24.0              # PDF text extraction (fitz)
pdf2image>=1.17.0            # PDF to image conversion
python-magic>=0.4.27         # File type detection

# ==========================================
# Audio Processing
# ==========================================
mutagen>=1.47.0                # Audio metadata extraction (duration, etc.)

# ==========================================
# Content Ingestion
# ==========================================
httpx>=0.26.0
aiohttp>=3.9.3
aiofiles>=23.2.1
beautifulsoup4>=4.12.3
lxml>=5.1.0
pypdf>=4.0.1
pdfplumber>=0.10.4
markdown>=3.5.2
python-frontmatter>=1.1.0
watchdog>=4.0.0
ebooklib>=0.18
youtube-transcript-api>=0.6.2
trafilatura>=1.6.0           # Web article extraction
# openai-whisper>=20231117  # Heavy dependency (torch) - install separately if needed

# ==========================================
# Knowledge Hub / Obsidian
# ==========================================
jinja2>=3.1.3
ruamel.yaml>=0.18.6
pyyaml>=6.0.1

# ==========================================
# Logging & Configuration
# ==========================================
structlog>=24.1.0
python-dotenv>=1.0.1
rich>=13.7.0

# ==========================================
# Data Processing
# ==========================================
numpy>=1.26.0
scipy>=1.12.0
orjson>=3.9.13
typing-extensions>=4.9.0
python-dateutil>=2.8.2
pytz>=2024.1

# ==========================================
# Development & Testing
# ==========================================
pytest>=8.0.0
pytest-asyncio>=0.23.4
pytest-cov>=4.1.0
pytest-mock>=3.12.0
factory-boy>=3.3.0
faker>=22.5.1
respx>=0.20.2
black>=24.1.1
ruff>=0.2.0
mypy>=1.8.0
ipython>=8.21.0
